fix news
connection helpers and rename controller (getNews -> newsController
